refactor(#18): JwtIssueHistoryResponse → JwtHistoryResponse にリネーム
